Allow option to update only one setting
Previously, we did not have an option for instructors to update only one
of the settings. This features adds checkboxes for instructors to
indicate what they would like to update, and implements data validation
in both browser and server to respond appropriately. We chose this
approach as opposed to having separate update buttons for each setting,
to reduce the amount of update time for when instructors would like to
change more than one setting simulatneously.