Add warning and link to course outline
Currently, we publish all public problems when they get bulk updated. It
may be that the problem already is a draft with other changes. We warn
the user to visit the Course Outline to ensure they want these changes
to also be made public before they perform a bulk update.