Fix ordering bug on CME report
I had made a last-minute review comment about not splitting the lookup
away from the debug message that announces it, unless there was reason
to be moving it. Turns out, the lookup _did_ need moved and the
appropriate action would have been to move the debug statement too.

The `--include-grades` flag fails without this change.